Pursuit of Phase I
is an implementation a proof-of-concept (POC) that someone can actually use and get a feel for the product. 

Therefore, I used JSONPlaceholder for "Fake Online REST API for Testing and Prototyping."

Three components and one service were created for interaction with Todo object with the following attributes: id: number, title: string, and competed: boolean:

1. TodoItemComponent 
This is the individual todo item, that displays the title of the todo item, as well as a marked as completed checkbox and delete button.

- This component inputs Todo models from the outer component TodoListComponent.

- This component outputs the deleteTodo function to the the outer component TodoListComponent.

2. TodoListComponent 
This is the list of all todo items, made up of TodoItemComponents. 

3. TodoNewComponent 
This is the single form entry at the top of the page for adding new todo items. 

- This component outputs the addTodo function to the the outer component TodoListComponent.

4. TodoService 
This service enacts create, update and delete operations on the fake backend, JSON placeholder. 

One route was created as well, that redirects to the TodoListComponent as it contains the main segment of the required functionality.